requirements.txt
alembic>=1.17.0
apscheduler>=3.10.0
asyncpg>=0.30.0
boto3>=1.35.0
fastapi>=0.119.0
httpx>=0.28.1
psycopg2-binary>=2.9.11
pydantic-settings>=2.11.0
pyjwt>=2.10.1
redis>=6.4.0
sqlalchemy>=2.0.44
sqlmodel>=0.0.27
uvicorn[standard]>=0.37.0
